+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release Note
 
+## v0.1.0
+
+- Added **Tweets Likes** endpoints. ([#6](https://github.com/twitter-dart/twitter-api-v2/issues/6))
+  - DELETE /2/users/:id/likes/:tweet_id
+  - GET /2/tweets/:id/liking_users
+  - GET /2/users/:id/liked_tweets
+  - POST /2/users/:id/likes
+
 ## v0.0.1
 
 - First Release